tests/wpt/tests/dom/nodes/getElementsByClassName-18.htm

<!doctype html>
<html class="a,b">
 <head>
  <title>element.getElementsByClassName(array): "a", "b"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a,b x">
  <div id="log"></div>
  <p id="r" class="a,bx"></p>
  <script class="xa,b">test(function() {
                             assert_array_equals(document.documentElement.getElementsByClassName(["\fa","b\n"]),
                                                 [document.body])
                            })
  </script>
 </body>
</html>
